app/requirements.txt
endee==0.1.10
google-generativeai
python-dotenv
requests
